services/hospital_node/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
torch==2.3.0
torchvision==0.18.0
opacus==1.4.0
grpcio==1.64.0
protobuf==5.27.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
cryptography==42.0.8
